dont crash if link dest already exists

+++ bin/utils.py
@@ -1143,7 +1143,8 @@ def copyFile(src, dest,linkOnly = envAsBool("EMERGE_USE_SYMLINKS")):
         if src.endswith(".exe") or src.endswith("qt.conf"):
             shutil.copy( src , dest )
         else:
-            os.symlink(deSubstPath(os.path.realpath(src)), dest )
+            if not (os.path.exists(dest) or os.path.islink(dest)):#dont collide with existing links
+                os.symlink(deSubstPath(os.path.realpath(src)), dest )
     else:
         shutil.copy(src,dest)
     return True
